check version

+++ tabulous/_qt/_table/_base/_enhanced_table.py
@@ -153,7 +153,7 @@ class _QTableViewEnhanced(QtW.QTableView):
         self._h_default = table.row_size
         self._w_default = table.column_size
         self._font = table.font
-        self.setFont(QtGui.QFont(self._font, self._zoom * self._font_size))
+        self.setFont(QtGui.QFont(self._font, int(self._zoom * self._font_size)))
 
     @property
     def _focused_widget(self) -> QtW.QWidget | None:
